**Only merge using a merge commit!**

+++ app/src/fleet/actions/deleteManufacturer.ts
@@ -1,8 +1,6 @@
 "use server";
 
-import { authenticateAction } from "@/auth/server";
-import { serverActionErrorHandler } from "@/common/actions/serverActionErrorHandler";
-import type { ServerAction } from "@/common/actions/types";
+import { createAuthenticatedAction } from "@/actions/utils/createAction";
 import { prisma } from "@/db";
 import { revalidatePath } from "next/cache";
 import { z } from "zod";
@@ -11,55 +9,41 @@ const schema = z.object({
   id: z.string().cuid(),
 });
 
-export const deleteManufacturerAction: ServerAction = async (formData) => {
-  try {
-    /**
-     * Authenticate and authorize the request
-     */
-    const authentication = await authenticateAction("deleteManufacturerAction");
-    await authentication.authorizeAction(
-      "manufacturersSeriesAndVariants",
-      "manage",
-    );
-
-    /**
-     * Validate the request
-     */
-    const { id } = schema.parse({
-      id: formData.get("id"),
-    });
+export const deleteManufacturer = createAuthenticatedAction(
+  "deleteManufacturer",
+  schema,
+  async (formData, authentication, data) => {
+    if (
+      !(await authentication.authorize(
+        "manufacturersSeriesAndVariants",
+        "manage",
+      ))
+    )
+      return {
+        error: "Du bist nicht berechtigt, diese Aktion auszuführen",
+        requestPayload: formData,
+      };
 
     /**
      * Delete
      */
     await prisma.manufacturer.delete({
       where: {
-        id,
+        id: data.id,
       },
     });
 
     /**
      * Revalidate cache(s)
      */
-    revalidatePath(`/app/fleet/settings`);
+    revalidatePath("/app/fleet/settings");
     revalidatePath("/app/fleet");
 
     /**
      * Respond with the result
      */
     return {
-      status: 200,
+      success: "Erfolgreich gelöscht",
     };
-  } catch (error) {
-    return serverActionErrorHandler(error, {
-      errorMessages: {
-        "400": "Ungültige Anfrage",
-        "401": "Du musst angemeldet sein, um diese Aktion auszuführen",
-        "403": "Du bist nicht berechtigt, diese Aktion auszuführen",
-        "404":
-          "Beim Löschen ist ein Fehler aufgetreten. Der Hersteller konnte nicht gefunden werden.",
-        "500": "Beim Löschen ist ein unerwarteter Fehler aufgetreten",
-      },
-    });
-  }
-};
+  },
+);
